Added clip/rotate pipeline.
