+++ main.py
@@ -1,18 +1,22 @@
 from os import listdir
 from os.path import isfile, join
 
+import fire
+
 from src import TEST_FOLDER
-from src.data.feature_matching_preprocessor import preprocess_data as feature_preprocess_data
+from src.data.preprocessor import preprocess_data
 
 
-def preprocess_test_data():
-    test_files = [f for f in listdir(TEST_FOLDER) if isfile(join(TEST_FOLDER, f))]
+def preprocess_test_data(file_name="."):
+    if file_name == ".":
+        test_files = [f for f in listdir(TEST_FOLDER) if isfile(join(TEST_FOLDER, f))]
+    else:
+        test_files = [file_name]
     for test_file_name in test_files:
         test_file_path = join(TEST_FOLDER, test_file_name)
-        print(test_file_path)
-        # preprocess_data(test_file_path)
-        feature_preprocess_data(test_file_path)
+        print(f'Processing {test_file_path}')
+        preprocess_data(test_file_path)
 
 
 if __name__ == "__main__":
-    preprocess_test_data()
+    fire.Fire(preprocess_test_data)
